LayoutTests/imported/w3c/web-platform-tests/intersection-observer/padding-clip-expected.txt


PASS Scrollport is used rather than content rect to compute intersection ratio
